Implement support for gain maps in OpenGL

+++ app/src/main/java/amirz/dngprocessor/parser/CFAPattern.java
@@ -7,24 +7,24 @@ import java.util.HashMap;
 import java.util.Map;
 
 public class CFAPattern {
-    private static final Map<int[], Integer> PATTERNS = new HashMap<>();
+    private static final Map<byte[], Integer> PATTERNS = new HashMap<>();
 
     static {
-        PATTERNS.put(new int[] { 0, 1, 1, 2 },
+        PATTERNS.put(new byte[] { 0, 1, 1, 2 },
                 CameraCharacteristics.SENSOR_INFO_COLOR_FILTER_ARRANGEMENT_RGGB);
 
-        PATTERNS.put(new int[] { 1, 0, 2, 1 },
+        PATTERNS.put(new byte[] { 1, 0, 2, 1 },
                 CameraCharacteristics.SENSOR_INFO_COLOR_FILTER_ARRANGEMENT_GRBG);
 
-        PATTERNS.put(new int[] { 1, 2, 0, 1 },
+        PATTERNS.put(new byte[] { 1, 2, 0, 1 },
                 CameraCharacteristics.SENSOR_INFO_COLOR_FILTER_ARRANGEMENT_GBRG);
 
-        PATTERNS.put(new int[] { 2, 1, 1, 0 },
+        PATTERNS.put(new byte[] { 2, 1, 1, 0 },
                 CameraCharacteristics.SENSOR_INFO_COLOR_FILTER_ARRANGEMENT_BGGR);
     }
 
-    public static int get(int[] cfaValues) {
-        for (Map.Entry<int[], Integer> kvp : PATTERNS.entrySet()) {
+    public static int get(byte[] cfaValues) {
+        for (Map.Entry<byte[], Integer> kvp : PATTERNS.entrySet()) {
             if (Arrays.equals(kvp.getKey(), cfaValues)) {
                 return kvp.getValue();
             }
